updates to HTML and endpoints to support better usage of random quote generator and documentation of avavailable endpoints. Creates endpoint to list movies
